Move random exam to ExamService
