+++ QuizManagement.WebApplication/Areas/Admin/Controllers/Exam/ExamController.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
@@ -87,7 +86,7 @@ namespace QuizManagement.WebApplication.Areas.Admin.Controllers.Exam
         }
 
         [HttpPost]
-        public IActionResult AddRandomQuestions(RandomExamModel model)
+        public IActionResult AddRandomQuestions(RandomExamViewModel model)
         {
             if (!ModelState.IsValid)
             {
@@ -95,60 +94,25 @@ namespace QuizManagement.WebApplication.Areas.Admin.Controllers.Exam
                 return new BadRequestObjectResult(allErrors);
             }
 
-            var allQuestions = _questionService.GetBySubject(model.SubjectId).ToArray();
-            bool[] checkQuestion = new bool[allQuestions.Length];
-            if (allQuestions.Length == 0)
+            try
             {
-                return new BadRequestObjectResult("Question not found");
-            }
-            if (allQuestions.Length < model.QuestionsNumber)
-            {
-                return new BadRequestObjectResult("Too many questions");
-            }
-
-            ExamViewModel examVm = new ExamViewModel
-            {
-                ExamName = model.ExamName,
-                Time = model.Time,
-                Examiner = model.Examiner,
-                DateCreated = model.DateCreated,
-                DateModified = model.DateModified,
-                Status = model.Status,
-                QuestionExamDetailViewModels = new List<QuestionExamDetailViewModel>()
-            };
-
-            Random rand = new Random();
-            int j = -1;
-            for (int i = 0; i < model.QuestionsNumber; i++)
-            {
-                while (true)
+                var examVm = _examService.CreateRandom(model);
+                if (examVm.Id == 0)
                 {
-                    j = rand.Next(allQuestions.Length);
-                    if (!checkQuestion[j])
-                    {
-                        checkQuestion[j] = !checkQuestion[j];
-                        break;
-                    }
+                    _examService.Create(examVm);
                 }
-                var question = new QuestionExamDetailViewModel
+                else
                 {
-                    ExamId = examVm.Id,
-                    QuestionId = allQuestions[j].Id
-                };
-                examVm.QuestionExamDetailViewModels.Add(question);
-            }
+                    _examService.Update(examVm);
+                }
 
-            if (examVm.Id == 0)
-            {
-                _examService.Create(examVm);
+                _examService.SaveChanges();
+                return new OkObjectResult(examVm);
             }
-            else
+            catch (Exception e)
             {
-                _examService.Update(examVm);
+                return new BadRequestObjectResult(new ModelError(e));
             }
-
-            _examService.SaveChanges();
-            return new OkObjectResult(examVm);
         }
     }
 }
